refactor(useTeamValues): Extract function to build team values
