+++ packages/client/src/modules/play/context/hooks/shared/useTeamValues.ts
@@ -1,7 +1,6 @@
 import { useMemo } from "react";
 import { TeamValues, usePersonaByUserId, usePlay } from "../../playContext";
-import { mean } from "../../../../../lib/math";
-import { buildStepToData } from "./utils";
+import { buildTeamValues } from "./utils";
 
 export { useTeamValues };
 
@@ -19,66 +18,12 @@ function useTeamValues(): {
 
   const teamValues = useMemo(() => {
     return teams.map((team) => {
-      const playersInTeam = players.filter((p) => p.teamId === team.id);
-      const playerRepresentingTeam = playersInTeam[0] || null;
-      const personaRepresentingTeam =
-        personaByUserId[playerRepresentingTeam?.userId || -1] || null;
-      const currentPersonaRepresentingTeam =
-        personaRepresentingTeam?.getPersonaAtStep?.(game.step) || null;
-
-      return {
-        id: team.id,
-        playerCount: playersInTeam.length,
-        points: mean(
-          playersInTeam.map(
-            ({ userId }) => personaByUserId[userId].currentPersona.points
-          )
-        ),
-        budget: mean(
-          playersInTeam.map(
-            ({ userId }) => personaByUserId[userId].currentPersona.budget
-          )
-        ),
-        budgetSpent: mean(
-          playersInTeam
-            .map(({ userId }) => personaByUserId[userId])
-            .map(
-              (persona) =>
-                persona.getPersonaAtStep(0).budget -
-                persona.currentPersona.budget
-            )
-        ),
-        carbonFootprint: mean(
-          playersInTeam.map(
-            ({ userId }) =>
-              personaByUserId[userId].currentPersona.carbonFootprint
-          )
-        ),
-        carbonFootprintReduction: mean(
-          playersInTeam
-            .map(({ userId }) => personaByUserId[userId])
-            .map(
-              (persona) =>
-                (1 -
-                  persona.currentPersona.carbonFootprint /
-                    persona.getPersonaAtStep(0).carbonFootprint) *
-                100
-            )
-        ),
-        stepToConsumption: buildStepToData(
-          "consumption",
-          game,
-          playersInTeam,
-          personaByUserId
-        ),
-        stepToProduction: buildStepToData(
-          "production",
-          game,
-          playersInTeam,
-          personaByUserId
-        ),
-        productionCurrent: currentPersonaRepresentingTeam?.production || [],
-      };
+      return buildTeamValues({
+        game,
+        personaByUserId,
+        players,
+        team,
+      });
     });
     // TODO: check `personaByUserId` in deps doesn't trigger infinite renders.
   }, [game, personaByUserId, players, teams]);
